Button Separated from App.js
